feat: propagate locale in SDK attribute layer (#196)
- [x] Requires a new release of Common (6.2.0?) before merging,
currently on Common's `main`

Plumbs the `CultureInfo.CurrentCulture` locale string into the SDK
layer.

The platform layer provides `null` since this isn't a cross-platform
thing and is available in standard .NET.

Caveat: I haven't tested this on Android or iOS. It's possible the
standard `CultureInfo.CurrentCulture` just doesn't work on those
platforms, or something, and we actually _do_ need to use special APIs.

Also note: if the culture isn't set, then .NET uses the
"InvariantCulture". In this case, we'll pass "en" as the `locale` string
because it's English-like without a specific region. This is less
confusing than omitting or using an empty string.